feat: use transition-right glyph for flip-device-face verb (#2400)
Replace IconFlip (Iconoir "flip" mirror triangles) with a new
IconTransitionRight component (Iconoir transition-right glyph: solid
rounded panel, arrow, outlined panel) for the flip-device-face verb.
The previous glyph read as a generic mirror/transform; the new one
reads as "swap this face for the other one".

Swaps both usages of the icon (VerbBar verb map and DeviceDetails verb
template) and the barrel export, then deletes the now-unused IconFlip
component. Visual-only change: the verb's aria-label/title are unchanged.

Closes #2400
